MB-30970 Unnest scan does not choose union scan when unnest predicate has OR clause
When an UNNEST is done and unnest predicate has an OR clause with subterms of the OR
clause corresponds to different indexes, then previously in spock the planner chooses
an union scan, however in vulcan and later a primary index scan is used. The reason
is that in vulcan due to the new expression infrastructure, predicates are now
separated for each keyspace, and unnest reference is treated as a different keyspace,
thus when planning is done for the primary keyspace, we don't see the unnest
unnest predicates (with the OR clause). Only when buildUnnestScan() is called, then
the predicates on the unnest keyspace is added to predicates on the primary keyspace
before index selection. There is currently no special handling of OR clause in
buildUnnestScan(), thus we don't consider and generate an union scan.

The fix is to mark a keyspace as primary unnest when gathering keyspace referenes
from the FROM clause (if it unnests from the primary term). Then before index
selection for the primary keyspace, we add all filters from all primary unnests
such that those are also considered (for unnest scan). This logic was previously
inside buildUnnestScan() itself, now it is done much earlier so proper OR clause
processing with unnest scan can occur as well.

Change-Id: Ic8cb8c899f3d25260ef5714f2bfe64bacda54936
Reviewed-on: http://review.couchbase.org/98600
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>