MB-30924 improve util.uuid
uuid() (which is used to set requests ids) can be quite slow because of

https://github.com/golang/go/issues/22614

essentially access to /dev/urandom has to be serialized, and when everyone is piling in,
the cost of assigning a request id is (measured) 3.5% of the whole execution time!
This change makes the access to /dev/urandom asynchronous, reading in 32k chunks (one read
every 4k requests), which eliminates the cost completely.

It has been used in anger, at a rate of 50k requests/second, and it has not misbehaved.

Change-Id: I47e524ef11344aadafff770444f214093fe6e008
Reviewed-on: http://review.couchbase.org/98891
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>