MB-31001 refresh buckets on bucket delete and errors that may require a new map
This is related to MB - 30967. Essentially, when nodes restart, N1QL may spam them with
tens of thousands of requests even before the buckets are available.

The N1QL side of things entails marking keyspaces for refresh if the delete notifiers marks them as dead
(so far no action would be taken), or if errors on KV operations suggests that the vbmap may be out
of date and the bucket needs reloading.

This works in two ways:

- for ad-hoc requests, the statement will be parsed and planned, and the refresh will mean that the
KeyspaceByName() call by the optimizer will return a reloaded bucket if needed.

- for prepared statements (or auto prepare), the verifyKeyspace() operation will provide an up to date
bucket at execution time

In both cases, the chances of getting to a stale bucket in the microsecond it takes from checking to
executing are rather slim, and anyway the fact will be picked up checking the error, and addressed at
the next KeyspaceByName() call.

Change-Id: I23322275a7c475c157458ed8d0c46ccf77eae172
Reviewed-on: http://review.couchbase.org/98962
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>