MB-31079 USE INDEX hint on source not working properly in MERGE statement
For a MERGE statement, since there is no WHERE clause for MERGE, the only
available predicate for index selection is the ON-clause predicates. In
case an INSERT action is not specified, the join between source and target
is an inner join, and we should be able to use the ON-clause predicate for
index selection on the source. This is currently not done.

The fix is to move the processing of ON-clause processing to before we
perform index selection on the source, such that we could utilize the ON
clause predicate for index selection for the source. We also need to set
pushableOnclause to be the ON-clause so index selection works properly.

Change-Id: I9e2428b074a7fa5d17939f40d2e1e393f0d101a6
Reviewed-on: http://review.couchbase.org/98974
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>