MB-31156 crash with CTE and multiple documents
The defect is misleading, in that it implies that the issue with CTE, while instead it's down
to value handling.
More specifically, 

WITH a AS (1) SELECT * FROM default will return

[
  {
    "a": 1,
    "default": {
      "f": 1
    }
  },
  {
    "a": 1,
    "default": {
      "f": 2
    }
  }
]

eg the projection operator will take * to mean not default.*, but every single value
defined in the query, including the CTE.
The underlying issue is
that the parent value (a), is recycled as soon as the first scope value (default.f: 1)
has been returned to the client.
As soon as the next document comes along, the projection operator tries to add the parent
to the json output document, but since it has been recycled, we crash.

The fix is to have annotated, scope and parsed values have a reference count that is incremented
by NewScopeValue() every time the parent is referenced.
Recycling only happens when the refcount goes to 0.

We have now to decide if select * has to be implicitly qualified as keyspace.* by the
projection operator, or the current behaviour is correct.




Change-Id: Id60d20240a8bee038e65e10547d0e37d37eb62f5
Reviewed-on: http://review.couchbase.org/99721
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>