MB-31336 terminate first scan only after current queue has been processed
As indicated in the defect analysis this is a race condition whereby the intersect scan
takes the end of a scan as having already processed all the keys from that scan, which is not
the case.

The fix is as follows: when a scan terminates and we notice that the first scan is the last standing,
take a count of the keys still in the channel.
Once all those keys have been processed and there's no keys in the intersect scan queue, then it is
safe to stop the first scan.

Change-Id: I93d0530acb73bc14c5311e57edae4c9e385ce9b8
Reviewed-on: http://review.couchbase.org/99872
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>