MB-31381 improve performance of metrics recording
At the weekend I was doing some testing, and I have found that on an 80 clients run,
if I set the account store to stub:, my end to end latency would go down from 2.9ms to 2.5ms
and my throughput up from 17k sec to 20k sec.

Code inspection has found that gometrics exclusively locks the registry every time we look
for a metric, and we look for metrics 14 times at the end of each request!
(to put it in numbers, the metric registry lock is exercised 240k times a second in my little
cluster, and 1.25M on the large perfrunner cluster!)

Since the metrics are set up once, and never changed, we might just as well cache them for the
purpose of maintaining them (every other use is unchanged), thereby avoiding the lock altogether.

Change-Id: I3cfdeae04ba24b9c633abcb22c012a1faab7562c
Reviewed-on: http://review.couchbase.org/99923
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>