MB-31423 Intermittent error with chained LETTING variable and GROUP BY
The previous fix for MB-19772 ranges through a mapping of letting variables,
the problem is when ranging from a map the order is not deterministic. We
need to range from the order the letting variables are defined.

The fix is to range the letting array itself instead of a map of the letting
mappings.

Change-Id: I992add94b51625aea79c121a19367fb5bc7b86d4
Reviewed-on: http://review.couchbase.org/100007
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>