MB-30837 serialize stream
So the stream operator (the one 'sending' the results to the client) would receive a value,
send it down a channel, where it would be received be the servicer, marshalled into a buffer
(which is not pooled) and copied into another before being passed to the http writer.
A fairly convoluted chain of events for something that the stream operator could have done itself.

I have changed the Output interface in execution/context.go to have a SetUp() method (to do any
action needed prior to sending results), a Result() method - which physically sends the results),
and CloseResults() takes any cleanup action.

The request Execute() method now merely prints the prefix, releases the stream operator and waits
for completion, and then prints the postfix.
I have left prefix and postfix with the servicer because that ensures we still have a meaningful
result should the execution pipeline die with a panic of some sort.

The stream operator calls Result() directly which marshals the document straight in the writer
buffer - flushing the buffer before hand if time has elapsed, and after if size has exceeded.

Result() will truncate the buffer back to pre writing length if the marshaling goes wrong,
so as to have consistent output.

I also cleaned the buffer pool to not pool buffers far bigger than the KeepAlive setting - before
the change large documents would cause the buffer to grow out of control, and we'd still pool it,
growing memory.

With this change we avoid one buffer and two channels per request (saves on GC load), one send,
one receive and one copy per item.
More importantly, now the execution time for the stream operator is the actual execution time for
sending the data down the wire, which is valuable diagnostic information.

Improvements: +3% throughput -1% latency with small (20 bytes) documents. +5%, -2% with 1KB documents.

With this, I have hit 21k/sec on Q1 on my 8cores cluster, up from 8k/sec on 5.0, ie +250% throughput

Change-Id: I62f6079c60b06563133d4f67934c437dadae8ed6
Reviewed-on: http://review.couchbase.org/100112
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>