MB-31470 Join/nest/index_join/index_nest not working properly when plan is from encoded_plan
For the keyspaceTerm under join/nest/index_join/index_nest the ON KEYS or ON KEY expression
is now under joinKeys field, however when a plan is unmarshalled from encoded_plan we still
put it on the keys field resulting in runtime issues. This issue only manifest itself when
the query plan is not found in the prepared statement cache and has to be decoded from
encoded plan itself.

The fix is to put the key expression at the right place.

Change-Id: Icdde6e95358bea9308205614f883a6b804d0bacc
Reviewed-on: http://review.couchbase.org/100194
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>