MB-31615 TRUE predicate not properly handled in where clause
When there is a TRUE predicate (either true or something like 1 = 1) in
the where clause, current logic incorrectly assumes the entire where
clause is TRUE. The entire where clause is only TRUE when there is no
other where clause predicates. Note FALSE predicate is not an issue
since the entire where clause is false if one subterm (broken on AND
boundary) is FALSE, regardless of whether other predicates exist.

The fix is to keep track of whether any non-constant predicate exists
in the where clause, and only assumes the entire where clause is TRUE
when there is no other non-constant predicates.

Change-Id: I2cfe43ba5fac0b16a6b7c582a677fa41fa942653
Reviewed-on: http://review.couchbase.org/100492
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>