MB-31629 CTE join uses NL and not hash
Most of the work here has been done by Bingjie - my initial attempt was overly complicated.
I have just checked that it was complete in pushing and popping withs, fixed a little issue,
and extended it to avoid CTE aliases in the from clause.

We save the CTE variable names in the formalizer, and in the from_expression formalizer, if we find
an existing variable, but it's one of the allowed with, with avoid marking the expression as
correlated.
We use the same approach to avoid having to have aliases against CTE variable names in expression
scans.

When we formalize subqueries we save and later restore the parent withs, so that we have a stack of
with definitions, and only the top applies.

Change-Id: I81716647452f27b0e5aa6a30e7c1619b9b2659ba
Reviewed-on: http://review.couchbase.org/100758
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>