MB-31848 improve servicers throughput
What I observed on iris is that while the throughput for Q1 is 115K,
switching to a simple 'select 1' would lower the throughput rather
than increase it.

Executing a fatal request (eg prepared statement missing) goes at 270k/sec.
I have reworked the servicers not to have channels but just a load gate (for
the ifmx people, much like PDQ gates).

If there is less that `servicers` requests running, just run the request from
the ServeHttp thread, if not, wait in a fifo queue until a slot is free.

The queue is simple for now - I want to make it better, possibly lockless, but
at this moment in time on iris I jump from 100k to 160k with this code.

Change-Id: I294449edf58e3d9418f85b185e74817f64569b67
Reviewed-on: http://review.couchbase.org/101295
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>