MB-31381 bottleneck in Vitals
Timers are a meter and a histogram (which is a sample) - so there's no point in locking timers on update
because the underlying structures are locked.
Meters would have an update of the snapshot at every update - that's quite wasteful.
With this change the only bottleneck remaining is samples (which is the biggest one)

Change-Id: I49bcba770e15b670816c1bb3c7eeeb4cdcf81516
Reviewed-on: http://review.couchbase.org/101601
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>