MB-31381 shard sample reservoir to avoid contention
the reservoir is now partitioned in a number (8 for now) of equal sized slices.
on update choose a slice in round robin.
on access use the complete reservoir.
the sample is augmented with an inUse boolean for the benefit of the whole reservoir,
as now this is filled in partitions and not contiguously. 

Change-Id: I9e5ad1d33a9834635b18a91b7b9ab6c8fde4e98f
Reviewed-on: http://review.couchbase.org/101876
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>