MB-31700 lower instantiation time
This lower contention on the exchange infrastructure by not allocating value exchanges that are not
needed, for example sequences, parallel or serialized operators.
Ideally, I would like to get rid of the double loop on sequence children (one in construction, one in
execution) as it does take away a little bit of throughput - but coalescing the two loops does
create lots of execution problems, so...

Re this improvement, the last thing that's needed is to pool frequently used operators like
scan_primary3 and scan_index3, scan_union, the 3 DML's, etc.

Change-Id: I3b66c091044616ea2e9a4a20b49e764da9a907a6
Reviewed-on: http://review.couchbase.org/101902
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>