MB-31848 improve servicers throughput
This lowers the contention on the servicers queue lock.
The lock is only used to atomically modify run and queue count, and once a position in the queue
is reserved, waiters and wakers are allowed to go in parallel.
In fact this is no longer a fifo, queue entries, once reserved, can be occupied / freed out of order,
and a waker doesn't even have to wait for the waiter to populate the entry - it just marks it
as ready to go.

Throughput goes from 165 to 181 for a select raw 1 on the large iris cluster - and I have verified
correctness both empirically and through code inspection.

The queue mutex is still a tiny bit of a bottleneck, but the atomic run and queue count manipulations
I have tried (using atomic.CASUint64) are actually worse than this.

Change-Id: I2ece0052fb04110f507c3d94a36582a2a6b67a7f
Reviewed-on: http://review.couchbase.org/102078
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>