MB-32140: do not replace named/positional parameter with value during subquery execution for prepared statements
The fix for MB-24871 introduced a mechanism for replacing "static" named/positional
parameters with their values to facilitate better index selection. This replacement
should not be performed for prepared statements. This is checked for top-level
statement, but for a (non-FROM clause) subquery its plan is being built at execution
time and we currently do not have a similar check for prepared statements, resulting
in a named parameter being incorrectly replaced by a value during execution, such
that subsequent executions end up using the stale value for the parameter.

The fix is to prevent parameter replacement for prepared statements during execution.

Change-Id: I4d4d632cbb52f6c929687309ac8f0e5b208417fc
Reviewed-on: http://review.couchbase.org/102269
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>