MB-31848 improve servicers throughput
The underlying issue with the latest performance drop is that NUMA servers can't really cope
with the contention on the runQueue mutexes.
We are going to make this change in steps to aid with the review process.
The first step is to remove the mutexes slice and do a single CAS in the add and remove operations.
The CAS will either succeed, in which case add will proceed to sleep and remove will just mark the entry
as ready to go, or it won't in which case add knows that it can proceed (the entry was marked as ready to go),
and remove will have to wake up the waiter.

In the next review we are going to lower contention on the runQueue mutex - I have to possibilities in
mind: one is to just use atomic operations have have the actors deal with the theoretical race conditions
(there's the theoretical possibility that the would be waiter queues itself after all the services have
gone away), the other is to be formally correct but shard the queues and steal waiters from other queues
if none are found in the local one. 


Change-Id: I6c3178f316560fbc7cf15110898175e4f71ac854
Reviewed-on: http://review.couchbase.org/102330
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Isha Kandaswamy <isha@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>