MB-27171 pool values
This isn't the best pool that I could come up with (I am not so keen on the double mutex, but a single mutex is a serious bottleneck),
but it's the best I have come up with so far.
It lifts our theoretical upper limit to 300k/s and I can now do a simple select raw 1 at 200k/s.
Q1 has gone up to 127k/s on iris.

Also - I only have pooled values where they sit in a single operator pipeline.
This excludes insert/update/delete/filter as they could sit under merge and use the same values in parallel,
and collect because we need to cache results, and the results array is sent down the pipeline.

Window functions I still have to have a look at, and so the whole of the expression package.

Change-Id: I38953c31533f8be5262487aca69a4db65740a36f
Reviewed-on: http://review.couchbase.org/102427
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>