MB-32158 slowdown on IPV6 Q1
I have built a diagnostic n1ql to see what requests are coming into triton compared to iris, and this is the result:

iris

_time=2018-12-13T06:13:54.690-08:00 _level=INFO _msg=processing args [n1ql-000001081529] 
_time=2018-12-13T06:13:54.690-08:00 _level=INFO _msg=processing scan_consistency not_bounded 
_time=2018-12-13T06:13:54.690-08:00 _level=INFO _msg=processing timeout 75000000us 
_time=2018-12-13T06:13:54.690-08:00 _level=INFO _msg=processing prepared [172.23.100.55:8091]eacced90-b24e-568d-9b40-4228b5aed3f9 
_time=2018-12-13T06:13:54.690-08:00 _level=INFO _msg=processing encoded_plan H4sIAAAAAAAA/6SS...
triton

_time=2018-12-13T06:43:22.796-08:00 _level=INFO _msg=processing args [n1ql-000001460381] 
_time=2018-12-13T06:43:22.796-08:00 _level=INFO _msg=processing scan_consistency not_bounded 
_time=2018-12-13T06:43:22.796-08:00 _level=INFO _msg=processing timeout 75000000us 
_time=2018-12-13T06:43:22.796-08:00 _level=INFO _msg=processing prepared eacced90-b24e-568d-9b40-4228b5aed3f9 
_time=2018-12-13T06:43:22.796-08:00 _level=INFO _msg=processing encoded_plan H4sIAAAAAAAA/6SS...

IE triton (which uses IPV6 and canonical host names rather than ip addresses) is not used distributed
prepared statements, which means it has not yet determined that it is part of the cluster.

This is what N1QL clustering (http://172.23.132.16:8093/admin/clusters/default/nodes) looks like:

[
    {
         "cluster":"default",
         "name":"triton-srv-06-ip6.perf.couchbase.com:8091",
         "queryEndpoint":"http://[triton-srv-06-ip6.perf.couchbase.com:8091]:8093/query/service",
         "adminEndpoint":"http://[triton-srv-06-ip6.perf.couchbase.com:8091]:8093/admin",
         "querySecure":"https://[triton-srv-06-ip6.perf.couchbase.com:8091]:18093/query/service",
         "adminSecure":"https://[triton-srv-06-ip6.perf.couchbase.com:8091]:18093/admin",
         "options":null
    }
]

Note the endpoints included the node name, and not the actual canonical address.
This is down to HostNameAndPort() trying to apply IPV6 rules to a canonical name, and failing to split the
management port from the name.

I can hear you ask - what's that got to do with a performance problem?
Simples! N1QL clustering can't find its own node in the node services because of the malformed name splitting
because it fails to split properly the local port from the local name

https://github.com/couchbase/query/blob/master/clustering/couchbase/clustering_cb.go#L90

This means that when the node tries to find its own node name, it gets from clustering a soft failure
of the sort "you are not in the cluster yet", try again next time, and for every request execution
it queries ns_server 3 times and locks the clustering lock.

With this change, the node name is cache and ns_server is never queried again.


Change-Id: I43f5880ac06961c66fc4a48de80d516cba960198
Reviewed-on: http://review.couchbase.org/102802
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>