MB-32354 index condition as part of ON clause is not properly covered
The problematic query has two ANSI JOINs, both using nested-loop join,
and both uses covering index scans under the nested-loop join. During
ANSI JOIN planning we need to perform covering transformation on the
ON-clause. When this covering transformation is done for the 1st join,
the filter covers portion generates new identifiers from parsing the
filter string. The issue with this is that identifiers are now marked
with a flag indicating whether it's a keyspace identifier. This flag
is lost when a new identifier is generated by parsing the string.
Subsequently when the 2nd join is being considered, we incorrectly
determined that the filter_covers expression cannot be covered by
the index due to lack of keyspace flag on the identifier.

This is not an issue if the index condition filter is put in the WHERE
clause instead of the ON clause, the reason being WHERE clause gets
covering transformation after all the joins are done, while ON clause
gets covering transformation as part of the join planning, thus
covering transformation for one join do affect the next join.

The fix is to properly mark newly generated keyspace identifiers
for filter_covers after the expression is generated from parsing.

Change-Id: Ib5e94455c6ea3f831be726f1c2632b417372bd18
Reviewed-on: http://review.couchbase.org/102875
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>