MB-32440 panic in server/server.go:handleRequest()
there's a rare race condition whereby if two requests try to add themselves to the queue at
the same time and both exceed the queue size, only the first will be caught.
the second will add to the queue and overwrite a perfectly good entry, corrupting it.
Reading the entry at release time will cause a panic.

Extend the queue size test to catch anything.


Change-Id: I194e7fabe924949e63be801f34752f9130e520b1
Reviewed-on: http://review.couchbase.org/103157
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>