MB-31383 add global function names storage and system keyspace
This change implements metakv storage for function definitions and system keyspace to access them.
As part of that it implements a resolver mechanism which given a stored function, identifies the
correct language and unmarshals the body correctly.
Has to be separate from functions package to avoid recursive referencing with inline / expression.

Also had to cleanup the function cache keyspace type name to avoid conflict with the functions keyspace. 

Change-Id: Ia1af4f339f2bcbf54b0af8ba5bc0bb851a4d3904
Reviewed-on: http://review.couchbase.org/103742
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>