MB-32284 improve hash joins latency
It seems that the regression with higher latency in hash joins in perfrunner is down to the improvements in result writing.
I don't have a way to run the test, but I have trimmed result writing as much as I could, including removing
needless mutexes, defers, fprintf, value copies and memory allocation.

On Q1 this improves throughput from 129k to 132k and latency from 3.5ms to 2.5ms, so it can't be that bad! 
Will ask Alex to rerun the test.

Change-Id: Ibbcf0555af1a4e4e1ddde85875b92363cc775041
Reviewed-on: http://review.couchbase.org/104271
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>