MB-32872 Projection alias not properly marked in ORDER BY clause for covering considerations
This bug is a regression from the fix for MB-30495.
In covering determinations we now rely on proper marking of identifiers
with appropriate flag as keyspace identifier, binding variable identifier
or projection alias identifier. The issue here is that currently the
marking of projection alias identifier only happens when there is a
GROUP BY clause present. In problematic query without a GROUP BY clause
we do not properly mark projection alias identifiers in the ORDER BY
clause causing issues for covering determination.

The fix is to properly mark projection alias identifiers in ORDER BY
clause in algebra.

Change-Id: Id61c5dcf863eb298c4e847811eb50c9f43727d32
Reviewed-on: http://review.couchbase.org/104398
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>