MB-32314 contention on annotated value pool
with Q13, the annotated value pool gets hit at a rate of

4 operators * 50 documents * 480 clients == 72000 lock requests at any one time, which it cannot
cope with.
The correct fix is to have at least the projection operators not allocate so many annotated values.
For now - measuring the load and avoiding the pool if too much contention is a decent enough compromise.

With this change Q13 goes up from 14.5k to 27.5k on iris.

Change-Id: Ica7b2f3ab589e3d698b0a4cb90969075c85fd29d
Reviewed-on: http://review.couchbase.org/104566
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>