MB-31417 invoke UDFs from anywhere in an expression
This implements stuff like select udf(blah).
I had to break even more parser / function /expression cycles which I could not do
without resorting to a functional pointer to construct function names (functions.Constructor)
which meant a) having an explicit function initialization (that's needed or functions/constructor/init()
is never called because since it is implicitly referenced, the linker doesn't even know it has to link it)
and b) changing the functions algebra interface.

I also had to rejig the formalizer because something like

create function func1(parm1) { select * from `travel-sample` where type=parm1 }

would mistakenly mark the subselect as correlated - as the with aliases would be zeroed at a new
subselect formalizer creation. This might be related to MB-31827 too.

Finally, I have created groundwork to pass request scope to UDFs and prepareds statements which
will come handy for the collections work.
Its is all marked as TODO with so,e blurb on scopes.

I have sought to error out when preparing a statement that references a non existent UDF to
maintain similar behaviour as when referencing an unknown predefined function (to avoid test
failures)

As a side note - I have to ignore the namespace in the multistore prepareds test because
depending on whether they run as FS or CS, the namespace is "dimestore" or "multistore",
and the test rig can't cope with multiple namespaces...

Change-Id: I86ec2572e6fa98b5f4e19cc37742a3ab6b8e5499
Reviewed-on: http://review.couchbase.org/104995
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>