MB-33171 avoid connection leak from go-couchbase to ns_server
so the issue here is that when ns_server is under stress, requests to /pools/default/nodeServices are slow
and go-couchbase keeps on opening more an more connections.

the reason why there are so many requests to nodeServices is that although the distributed code caches nodes,
it uses the clustering code to obtain cluster maps in order to know which nodes to contact, and these are not
cached.

This was supposed to be fixed as part of MB 24229 (which requires a good rewrite of clustering), and MB 30992
was an attempt to improve things.

It turns out that there's no getting away from caching nodes and maps in clustering, so I have rewritten
system_remote.go to use as much cashable data as possible, and actually implemented some cache in clustering
(although it does not address the whole n1ql node map issue).
I have also undone MB-30992. With this change, I don't see excessive connections on the user's scenario.

Change-Id: Ie2c099730117b978dd2d740fa9d8cc110ebffea1
Reviewed-on: http://review.couchbase.org/105594
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>