MB-30839 resolve parser conflicts for full function names
The issue here is that there would be 2 shift / reduce conflicts when we would allow the use
of full function names (namespace:function() or namespace:bucket.scope.function()) in expression.

These were due to COLON being used as an operator in multiple places in expression and expressions
being used along side keyspace terms in the simple from clause.

The first was fairly easy to fix
essentially have 1 rule for namespace_name which includes the COLON (same thing for namespace_term
and SYSTEM), since we look ahead one token, seeing IDENT COLON or SYSTEM COLON means that we can
reduce to a namespace straight away.

The second is way more complicated to fix - seeing IDENT COLON IDENT, are we going to shift for the
function name or reduce for a slice?

I have battled with this for a few days, and in the end I had to resort to the age old technique of
scanner hacks:
Look ahead one token in the scanner, and if I see IDENT COLON and IDENT matches one of the namespaces
then return the special token NAMESPACE_ID.

This means that the parser has to be informed of the current namespaces, which currently are created
at store creation time, and never modified and so we can do it once and for all at server
initialization.

Not clean, but it works. 

Change-Id: I4fadf485f923c299e1ca2602a92dc40eabfb1ace
Reviewed-on: http://review.couchbase.org/105934
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>