MB-29292 Add 'hint_not_followed' to ansi join operators in explain
Hash join is only supported in enterprise edition, in community edition a hash join
hint is silently ignored.

Add a 'hint_not_followed' field in the explain plan for ansi join operators that
specifically indicates that hash join is not supported in community edition.

The UI team will need to take advantage of this new information and display
warnings appropriately as requested in the MB.

Change-Id: Id509290405ec60a982a613feb8902348cc6829bf
Reviewed-on: http://review.couchbase.org/106663
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>