MB-33374 pass security configuration to indexers
1. We maintain a ConnectionSecurityConfig structure to keep
track of TLS settings and node to node security settings.
2. ConnectionSecurityConfig is maintained by a callback
from cbauth. It is initially set by the callback very early
in the life of the process.
3. When the configuration changes, it is are passed to the datastore.
The datastore passes it on the the indexers, and they do whatever
updates they have to.
4. If the datastore has not received a ConnectionSecurityConfig,
it refuses to start up buckets, because buckets need indexers,
and indexers should not be started without knowing their
initial security configuration.

I tested this code in a couple of scenarios, and the query engine
comes up correctly in normal operation.

Change-Id: I55cc82d98d6912e0c1f899274e0671e2fa5ce723
Reviewed-on: http://review.couchbase.org/106864
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Tested-by: Johan Larson <johan.larson@couchbase.com>