MB-33928 Properly identify source of a MERGE statement to be a keyspace
In a select statement an identifier in the FROM clause is first treated as an expression,
and during formalization process it is then determined whether the identifier is
a keyspace reference or not. Subsequent handling of the expression term is based
on the isKeyspace indicator.

For the source of a MERGE statement, it is similarly created as an expression term,
however subsequent handling does not always check whether the expression is
actually a keyspace.

This creates issues for the planner when a keyspace name is required, e.g. when
looking for a matching histogram.

The fix is to check after formalization to see whether a keyspace is being used,
and if so, swtich the source to a keyspace term so subsequent handling treats it
as a keyspace.

Change-Id: If7aa85b3a3671e8ef6faef38803a1b4bddec65d4
Reviewed-on: http://review.couchbase.org/108233
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>