MB-33892 race condition in execution
The underlying issue is that killed operators where still expected to do cleanup and signaling to other operators, and essentially writing on base
even when they weren't supposed to be used.
	
This has been the case since the beginning of time.
Changed inactive operators to bail out straight away, and left the onus of cleanup on the Done() and SendStop() infrastructure.

Also avoid pooling operators that have been killed, as goroutines are still around and may read base
and act incorrectly on it.
	
Also added a little method to release waiters in the Recover() code to be used as part of MB - 31641

Change-Id: Ia2604d52284ac23c64a934ee1b9a4b18c7853cbb
Reviewed-on: http://review.couchbase.org/108669
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>