MB-33509 variables not pushed down in inline functions
Although this MB has been logged against functions, the problem is more general:
Consider this query:

select * from `travel-sample` where type = "hotel"

This will do a scan on the type index with a span using "hotel" as low and high.
On the other hand

with mytype as ("hotel")  select * from `travel-sample` where type = mytype

uses a value scan (which takes 10 times as long).
The underlying issue is that identifiers don't produce Static() expressions, so planner.getSarg() returns a nil
sargable expression.

We mark bindings for WITH expressions and function parameter lists as able to produce static values,
and propagate that to identifiers.
All other variables are untouched and won't produce a static expressions, and therefore won't be pushed down.


Change-Id: I94c41cf82d0720d8319ce6e3073b1acde066b9fe
Reviewed-on: http://review.couchbase.org/110156
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>