MB-34544 generate a special error for subquery build failures and set HTTP error code accordingly
Our system has a distinctive way of handling subqueries. We treat them as expressions
and defer building their plans until execution time. This means we encounter build problems
either right up front, for problems with the main query and somewhat later, for subqueries.
This also means the two have somewhat different error behavior.

The deep fix here would be to handle subquery plan generation right up front, when the plan
for the main query is prepared. That way, both sorts of plan problems would be found
at the same time. Unfortunately that would be a substantial architectural shift,
really not the sort of thing we should be trying this late in Mad-Hatter.

This CL attempts a narrower fix: we generate a special error for subquery build
problems, and enter it directly into the context, so we don't have to worry about
whatever the surrounding query or queries will do with it. We can then look for this
particular error code when the produce the result, and set the HTTP error code based
on it.

This does mean that the result contains two errors: one entered directly from the subquery,
and one produced from the error return to the surrounding query.

Change-Id: Ifad8b807fcb6c6649e229e36d2f26dd66c15c579
Reviewed-on: http://review.couchbase.org/110544
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Johan Larson <johan.larson@couchbase.com>