MB-33674 make sure that the stop signal is propagated
After having gone back and forth through loads of stacks,
the issue boils down to the clients disconnecting and the
engine handling the closed connection, propagating a stop
and waiting for the operators to go away.
This defensive fix makes sure that a stop signal is sent to
each exchange infrastructure.

Change-Id: I6af7448f5c13fbfcb4a15ae86a2dc51b554c4e42
Reviewed-on: http://review.couchbase.org/110887
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>