MB-34749 random empty results on prepared with covered subquery
This is a sideways fix, but it does have advantages!
We are making the subquery plan a property of the prepared, meaning that once we plan the first time,
we don't have to plan over and over again.
Since we don't replan, we don't get it wrong with the addition of the cover node in the algebra.

Advantages - first, we could actually explain subquery plans part of meta().plan for system:prepareds,
once we decide to implement profiling for subqueries.
Second, with luck, we should cut down on planning time for heavily used prepareds, meaning better throughput and latency. 

Change-Id: I076c3c5315f87b24e41571a74736c0f23dfb121c
Reviewed-on: http://review.couchbase.org/111713
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>